09.06.2022 MenuGame and animation scene home1
The main menu and buttons (Continue, New Game, Load Game, Exit) are fully implemented visually.
The script implemented buttons (Stop and Continue Game, New Game, Exit). Not made the function of the button save and load the game. Also, the rest of the made buttons do not work very well, we need to finish the little things!